feat: Enhance service details and contact map functionality
Adds an optional `details` field to the `Service` type for extended descriptions, displayed in a modal on the Services page. The Contact page now embeds a Google Map using a `mapUrl` from the `ContactData` type, replacing the placeholder image. Service items in the store have been updated with new `details` content. The Admin page includes an Accordion component and functionality to add/delete services.